--- hlx_statics/blocks/combined-carousel/combined-carousel.css
@@ -1,163 +1,0 @@
-main div.combined-carousel-container:has(.combined-carousel.background-color-white) {
-  background-color: white !important;
-}
-
-div.combined-carousel picture {
-  width: 100%;
-}
-
-div.combined-carousel picture img {
-  object-fit: cover;
-  height: auto;
-  width: 100%;
-}
-
-@media screen and (max-width: 700px) {
-  main div.combined-carousel div.carousel-container {
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-  }
-
-  main div.combined-carousel div.media-container,
-  main div.combined-carousel div.text-container {
-    width: 100% !important;
-    max-width: 100%;
-    padding: 16px;
-    text-align: center;
-  }
-
-  main div.combined-carousel .slide-arrow {
-    visibility: hidden;
-  }
-}
-main div.combined-carousel p.image-container picture {
-  left: 0;
-  width: 100%;
-  height: 0;
-  position: relative;
-  padding-bottom: 56.25%;
-}
-
-main div.combined-carousel {
-  max-width: 1280px;
-  margin: auto;
-}
-
-main div.combined-carousel .block-container {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-content: space-between;
-  position: relative;
-}
-
-main div.combined-carousel .carousel-heading {
-  margin-top: 1 !important;
-  margin-bottom: 8px !important;
-}
-
-main div.combined-carousel .carousel-container {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-content: space-between;
-}
-
-main div.combined-carousel .media-container {
-  flex: 0 1 50%;
-  min-height: 100%;
-  max-width: 50%;
-  padding: 0 8px;
-}
-
-main div.combined-carousel .text-container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  line-height: 1.6;
-  flex: 0 1 50%;
-  min-height: 100%;
-  max-width: 50%;
-  padding: 0 32px;
-}
-
-main div.combined-carousel .carousel-button-container {
-  display: flex;
-  flex-direction: row !important;
-  margin-top: 8px;
-  gap: 16px;
-}
-
-main div.combined-carousel .slider-wrapper {
-  position: relative;
-  overflow: hidden;
-  transition-property: transform;
-  box-sizing: content-box;
-  align-items: center;
-}
-
-main div.combined-carousel .slides-container {
-  width: 100%;
-  display: flex;
-  list-style: none;
-  margin: 1;
-  padding: 0;
-  margin-bottom: 1;
-  margin-top: 1;
-  overflow: hidden;
-  scroll-behavior: smooth;
-}
-
-main div.combined-carousel .slide {
-  width: 100%;
-  height: 100%;
-  flex: 1 0 100%;
-}
-
-main div.combined-carousel .slide-arrow {
-  margin: auto;
-  background-color: transparent;
-  border: none;
-  font-size: 5rem;
-  color: #005bb5;
-  cursor: pointer;
-}
-
-main .combined-carousel .slide-arrow:hover,
-main .combined-carousel .slide-arrow:focus {
-  color: #003e7a;
-}
-
-main div.combined-carousel .slides-container {
-  scrollbar-width: none; /* Firefox */
-  -ms-overflow-style: none; /* Internet Explorer 10+ */
-}
-
-/* WebKit */
-main div.combined-carousel .slides-container::-webkit-scrollbar {
-  width: 0;
-  height: 0;
-}
-
-main div.combined-carousel .carousel-circle-div {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  gap: 8px;
-}
-
-main div.combined-carousel .carousel-circle {
-  background-color: #c6c6c6;
-  border: none;
-  border-radius: 100%;
-  width: 12px;
-  height: 12px;
-  margin: 4px 2px;
-  cursor: pointer;
-}
-
-main div.combined-carousel .carousel-circle-selected {
-  background-color: #007aff !important;
-}
