--- hlx_statics/blocks/video-carousel/video-carousel.css
@@ -1,163 +1,0 @@
-div.video-carousel picture {
-    width: 100%;
-    height: 100%;
-    display: flex;
-  }
-  
-  div.video-carousel picture img {
-    object-fit: cover;
-    height: auto;
-    width: 100%;
-  }
-  
-  main div.video-carousel .image-container {
-    height: 300px;
-    max-width: 602px;
-  }
-  
-  main div.video-carousel {
-    max-width: 1280px;
-    margin: auto;
-  }
-  
-  main div.video-carousel .video-block-container {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-content: space-between;
-    position: relative;
-  }
-  
-  main div.video-carousel .video-carousel-heading {
-    margin-top: 1 !important;
-    margin-bottom: 8px !important;
-  }
-  
-  main div.video-carousel .video-container {
-    display: flex;
-    flex-direction: row-reverse;
-    justify-content: center;
-    align-content: space-between;
-  }
-  
-  main div.video-carousel .left-container {
-    flex: 0 1 50%;
-    min-height: 100%;
-    max-width: 50%;
-    padding: 0 8px;
-  }
-  
-  main div.video-carousel .right-container {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    line-height: 1.6;
-    flex: 0 1 50%;
-    min-height: 100%;
-    max-width: 50%;
-    padding: 0 32px;
-  }
-  @media screen and (max-width: 900px) {
-    main div.video-carousel div.video-container {
-      flex-direction: column;
-      align-items: center;
-      justify-content: center;
-    }
-  
-    main div.video-carousel div.left-container,
-    main div.video-carousel div.right-container {
-      width: 100% !important;
-      max-width: 100%;
-      padding: 16px;
-      text-align: center;
-    }
-  
-    main div.video-carousel .video-slide-arrow {
-      visibility: hidden;
-    }
-  }
-  main div.video-carousel .video-carousel-button-container {
-    display: flex;
-    flex-direction: row !important;
-    margin-top: 8px;
-    gap: 16px;
-  }
-  
-  main div.video-carousel .video-slider-wrapper {
-    position: relative;
-    overflow: hidden;
-    transition-property: transform;
-    box-sizing: content-box;
-    align-items: center;
-  }
-  
-  main div.video-carousel .video-slides-container {
-    width: 100%;
-    display: flex;
-    list-style: none;
-    margin: 1;
-    padding: 0;
-    margin-bottom: 1;
-    margin-top: 1;
-    overflow: hidden;
-    scroll-behavior: smooth;
-  }
-  
-  main div.video-carousel .video-slide {
-    width: 100%;
-    height: 100%;
-    flex: 1 0 100%;
-  }
-  
-  main div.video-carousel .embed {
-    margin-top: 0px;
-  }
-  
-  main div.video-carousel .video-slide-arrow {
-    margin: auto;
-    background-color: transparent;
-    border: none;
-    font-size: 5rem;
-    color: #005bb5;
-    cursor: pointer;
-  }
-  
-  main .video-carousel .video-slide-arrow:hover,
-  main .video-carousel .video-slide-arrow:focus {
-    color: #003e7a;
-  }
-  
-  main div.video-carousel .video-slides-container {
-    scrollbar-width: none; /* Firefox */
-    -ms-overflow-style: none; /* Internet Explorer 10+ */
-  }
-  
-  /* WebKit */
-  main div.video-carousel .video-slides-container::-webkit-scrollbar {
-    width: 0;
-    height: 0;
-  }
-  
-  main div.video-carousel .video-carousel-circle-div {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    gap: 8px;
-  }
-  
-  main div.video-carousel .video-carousel-circle {
-    background-color: #c6c6c6;
-    border: none;
-    border-radius: 100%;
-    width: 12px;
-    height: 12px;
-    margin: 4px 2px;
-    cursor: pointer;
-    font-size: 24px;
-  }
-  
-  main div.video-carousel .video-carousel-circle-selected {
-    background-color: #007aff !important;
-  }
-
